requirements.txt
Django>=5.0
requests[socks]
PySocks
lxml
Pillow
python-decouple
